add: splitting code  into multilayers - createTodo
